Getting nuts codes
