+++ services/enrichment/manager/src/lib/enrich.js
@@ -1,22 +1,38 @@
 import { mergeRecords } from './merge';
-import { enrichLocationFromAddress } from './enrichLocationFromAddress';
-import { enrichLocationFromCentroid } from './enrichLocationFromCentroid';
+import { enrichFromCoordinates } from './plugins/coordinates';
+import { enrichFromCountry } from './plugins/country';
 
 export const enrich = async (record, existingRecord) => {
   const enrichedRecord = mergeRecords(record, existingRecord);
 
   const newLocations = (await Promise.all(
     enrichedRecord.project_locations.map(async loc => {
+      // No information to work with, just exit
       if (!loc) {
         return null;
       }
 
-      if (!loc.centroid && loc.country_code) {
-        // Centroid is empty but country code is provided
-        return enrichLocationFromAddress(loc);
-      } else if (!loc.country_code && loc.centroid) {
-        // Country code is empty but centroid is provided
-        return enrichLocationFromCentroid(loc);
+      // Best scenario: we have coordinates
+      if (loc.centroid) {
+        // get nuts information from coordinates
+        enrichFromCoordinates(loc);
+      } else if (loc.address) {
+        // get coordinates from address
+        // if data added, call previous enrichFromCoordinates plugin here as well
+      }
+
+      // Less precision
+      if (loc.nuts.length) {
+        // get country code if not available yet
+        // separate or only in previous step?
+      }
+
+      if (loc.country_code) {
+        // find centroid's coordinates and add them
+        enrichFromCountry(loc);
+        // marked as enriched in order to avoid further enrichment based on low precision
+        // eslint-disable-next-line
+        loc.enriched = true;
       }
 
       return loc;
