+++ services/enrichment/manager/src/lib/plugins/README.md
@@ -1,0 +1,3 @@
+# Enrichment plugins
+
+The name of each plugin points to the information source used for enrichment procedure.
