+++ services/enrichment/manager/src/lib/plugins/country.js
@@ -1,7 +1,7 @@
 import countries from 'i18n-iso-countries';
 import request from 'request-promise-native';
 
-export const enrichLocationFromAddress = async loc => {
+export const enrichFromCountry = async loc => {
   // Get country name from ISO 3166-1 Alpha-2 code (doesn't seem mandatory after some testing...)
   const country = countries.getName(loc.country_code, 'en');
 
@@ -57,4 +57,4 @@ export const enrichLocationFromAddress = async loc => {
   return loc; // location not enriched
 };
 
-export default enrichLocationFromAddress;
+export default enrichFromCountry;
